refactored logic into using subreconcilers
