+++ cmd/main.go
@@ -19,6 +19,7 @@ package main
 import (
 	"flag"
 	"github.com/sinamna/BasicAthenticator/internal/config"
+	"github.com/sinamna/BasicAthenticator/internal/controller/basic_authenticator"
 	"os"
 
 	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
@@ -33,7 +34,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
 	authenticatorv1alpha1 "github.com/sinamna/BasicAthenticator/api/v1alpha1"
-	"github.com/sinamna/BasicAthenticator/internal/controller"
 	//+kubebuilder:scaffold:imports
 )
 
@@ -95,7 +95,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&controller.BasicAuthenticatorReconciler{
+	if err = (&basic_authenticator.BasicAuthenticatorReconciler{
 		Client:       mgr.GetClient(),
 		Scheme:       mgr.GetScheme(),
 		CustomConfig: customConfig,
