--- internal/controller/const.go
@@ -1,6 +1,0 @@
-package controller
-
-const (
-	nginxDefaultImageAddress  = "nginx:1.25.3"
-	nginxDefaultContainerName = "nginx"
-)
